requirements.txt
# Mini Researcher Agent - PocketFlow Edition
# ============================================================================
#
# Installation:
#   pip install -r requirements.txt
#   # OR with uv:
#   uv sync
#
# Note: Tectonic (LaTeX engine) must be installed separately
#   curl --proto '=https' --tlsv1.2 -fsSL https://drop-sh.fullyjustified.net | sh
# ============================================================================

# ============================================================================
# Core Framework (replaces smolagents + mini-swe-agent)
# ============================================================================
pocketflow>=0.0.3           # 100-line LLM framework for workflow orchestration
python-dotenv>=1.0.0        # Environment variable management

# ============================================================================
# LLM & API Integration
# ============================================================================
openai>=1.0.0               # OpenRouter uses OpenAI SDK format
requests>=2.31.0            # HTTP requests for APIs
pydantic>=2.0.0             # Structured data validation

# ============================================================================
# Experimentation (Docker support)
# ============================================================================
docker>=7.0.0               # Optional: For isolated experiments (ENABLE_DOCKER=True)

# ============================================================================
# Scientific Skills Dependencies
# ============================================================================

# Literature Survey Skill
arxiv>=2.1.0                # arXiv API client
beautifulsoup4>=4.12.0      # HTML parsing for web scraping
bibtexparser>=1.4.0         # BibTeX file parsing and generation

# Method Implementation Skill
gitpython>=3.1.0            # Git provenance tracking

# Experimental Evaluation Skill
numpy>=1.24.0               # Numerical computing
pandas>=2.0.0               # Data manipulation
matplotlib>=3.7.0           # Plotting
scipy>=1.10.0               # Scientific computing
scikit-learn>=1.3.0         # Machine learning utilities
seaborn>=0.13.0             # Statistical data visualization

# Paper Writing Skill
PyYAML>=6.0                 # YAML configuration parsing
# Note: Tectonic binary required for PDF compilation

# ============================================================================
# Configuration & Utilities
# ============================================================================
# All scientific-skills dependencies included above
# No smolagents, mini-swe-agent, litellm, or tavily-python needed

# Estimated total size: ~50MB (vs 170MB previous)
